ccass-conversations/pyproject.toml
[project]
name = "ccass-conversations"
version = "0.1.0"
description = "CCaSS Conversations - Retrieve and process conversations from Microsoft Dataverse"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "msal>=1.30",
  "python-dotenv>=1.0",
  "pandas>=2.0",
]

[project.scripts]
ccass-conversations = "ccass_conversations.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
